requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
Pygments==2.17.2
python-magic==0.4.27
Werkzeug==3.0.1
MarkupSafe==2.1.5
Jinja2==3.1.3
click==8.1.7
itsdangerous==2.1.2
protobuf==4.25.2